Enable warning-free compilation even with -Wextra.

+++ src/expr.c
@@ -798,7 +798,7 @@ static int Add(void)
 	v3.type = STR_TYPE;
 	l1 = strlen(v1.v.str);
 	l2 = strlen(v2.v.str);
-	if (MaxStringLen && (l1 + l2 > MaxStringLen)) {
+	if (MaxStringLen && (l1 + l2 > (size_t) MaxStringLen)) {
 	    DestroyValue(v1); DestroyValue(v2);
 	    return E_STRING_TOO_LONG;
 	}
